python/requirements.txt
fastapi
pytest
uvicorn
pika
psycopg2-binary